src/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
mssql-django==1.5
pyodbc==5.1.0
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
